fix setting StartUIDelegate
